Split apart normalize_cfg.rs, taking all tests with it
