+++ hugr-passes/src/normalize_cfg.rs
@@ -1,0 +1,879 @@
+//! Merge BBs along control-flow edges where the source BB has no other successors
+//! and the target BB has no other predecessors.
+use std::collections::HashMap;
+
+use hugr_core::extension::prelude::UnpackTuple;
+use hugr_core::hugr::views::RootChecked;
+use hugr_core::hugr::{HugrError, hugrmut::HugrMut};
+use hugr_core::types::{EdgeKind, Signature, TypeRow};
+use itertools::Itertools;
+
+use hugr_core::ops::handle::CfgID;
+use hugr_core::ops::{CFG, DFG, DataflowParent, ExitBlock, OpTag, OpType};
+use hugr_core::{Direction, HugrView, OutgoingPort, PortIndex};
+
+use crate::ComposablePass;
+#[expect(deprecated)] // refactor/inline merge_basic_blocks, or just hide it
+use crate::merge_bbs::merge_basic_blocks;
+
+/// Errors from [normalize_cfg]
+#[derive(Clone, Debug, thiserror::Error, PartialEq, Eq)]
+pub enum NormalizeCFGError {
+    /// The requested node was not a CFG
+    #[error("Requested node was not a CFG but {_0}")]
+    NotCFG(OpTag),
+}
+
+/// Result from [normalize_cfg], i.e. a report of what changes were made to the Hugr.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum NormalizeCFGResult {
+    /// The entire [`CFG`] was converted into a [`DFG`].
+    ///
+    /// The entrypoint node id is preserved after the conversion, but now identifies
+    /// the new [`DFG`].
+    CFGToDFG,
+    /// The CFG was preserved, but the entry or exit blocks may have changed.
+    #[allow(missing_docs)]
+    CFGPreserved {
+        entry_changed: bool,
+        exit_changed: bool,
+    },
+}
+
+/// A [ComposablePass] that normalizes CFGs (i.e. [normalize_cfg]) in a Hugr.
+#[derive(Clone, Debug)]
+pub struct NormalizeCFGPass<N> {
+    cfgs: Vec<N>,
+}
+
+impl<N> Default for NormalizeCFGPass<N> {
+    fn default() -> Self {
+        Self { cfgs: vec![] }
+    }
+}
+
+impl<N> NormalizeCFGPass<N> {
+    /// Allows mutating the set of CFG nodes that will be normalized.
+    ///
+    /// If empty (the default), all (non-strict) descendants of the [HugrView::entrypoint]
+    /// will be normalized.
+    pub fn cfgs(&mut self) -> &mut Vec<N> {
+        &mut self.cfgs
+    }
+}
+
+impl<H: HugrMut> ComposablePass<H> for NormalizeCFGPass<H::Node> {
+    type Error = NormalizeCFGError;
+
+    /// For each CFG node that was normalized, the [NormalizeCFGResult] for that CFG
+    type Result = HashMap<H::Node, NormalizeCFGResult>;
+
+    fn run(&self, hugr: &mut H) -> Result<Self::Result, Self::Error> {
+        let cfgs = if self.cfgs.is_empty() {
+            let mut v = hugr
+                .entry_descendants()
+                .filter(|n| hugr.get_optype(*n).is_cfg())
+                .collect::<Vec<_>>();
+            // Process inner CFGs first, in case they are removed
+            // (if they are in an unreachable block when the Entry node has only the Exit as successor)
+            v.reverse();
+            v
+        } else {
+            self.cfgs.clone()
+        };
+        let mut results = HashMap::new();
+        for cfg in cfgs {
+            let res = normalize_cfg(&mut hugr.with_entrypoint_mut(cfg))?;
+            results.insert(cfg, res);
+        }
+        Ok(results)
+    }
+}
+
+/// Normalize a CFG in a Hugr:
+/// * Merge consecutive basic blocks i.e. where a BB has only a single successor which
+///   has no predecessors
+/// * If the entry block has only one successor, and no predecessors, then move its contents
+///   outside/before CFG.
+/// * (Similarly) if the exit block has only one predecessor, then move contents
+///   outside/after CFG.
+///    * If that predecessor is the entry block, then convert the CFG to a DFG.
+///      This will be reported via [NormalizeCFGResult::CFGToDFG]
+///
+///
+/// # Errors
+///
+/// [NormalizeCFGError::NotCFG] If the entrypoint is not a CFG
+#[expect(deprecated)] // inline/combine/refactor with merge_bbs, or just hide latter
+pub fn normalize_cfg<H: HugrMut>(mut h: &mut H) -> Result<NormalizeCFGResult, NormalizeCFGError> {
+    let checked: RootChecked<_, CfgID<H::Node>> = RootChecked::<_, CfgID<H::Node>>::try_new(&mut h)
+        .map_err(|e| match e {
+            HugrError::InvalidTag { actual, .. } => NormalizeCFGError::NotCFG(actual),
+            _ => unreachable!(),
+        })?;
+    merge_basic_blocks(checked);
+    let cfg_node = h.entrypoint();
+
+    // Further normalizations with effects outside the CFG
+    let [entry, exit] = h.children(cfg_node).take(2).collect_array().unwrap();
+    let entry_blk = h.get_optype(entry).as_dataflow_block().unwrap();
+    let cfg_parent = h.get_parent(cfg_node).unwrap();
+    // 1. If the entry block has only one successor, and no predecessors, then move contents
+    // outside/before CFG.
+    // (Note if the entry block has multiple successors, but no predecessors, we *could* move
+    // its contents outside/before the CFG, but would need to keep an empty/identity entry
+    // block - we do not do this here.)
+    let mut entry_changed = false;
+    #[allow(clippy::match_result_ok)] // let Ok(...) without .ok() borrows `h`
+    if let Some(succ) = h.output_neighbours(entry).exactly_one().ok() {
+        if succ == exit {
+            // Any (basic-block) predecessors of `entry` are unreachable, so allow/ignore
+            assert_eq!(
+                &Signature::new(
+                    entry_blk.inputs.clone(),
+                    entry_blk.successor_input(0).unwrap()
+                ),
+                h.signature(cfg_node).unwrap().as_ref()
+            );
+            // 1a. Turn the CFG into a DFG containing only what was in the entry block
+            // Annoying here - "while let Some(blk) = cfg.children(...).skip(1).next()" keeps iterator alive
+            let children_to_remove: Vec<_> = h.children(cfg_node).skip(1).collect();
+            for blk in children_to_remove {
+                h.remove_subtree(blk);
+            }
+            while let Some(ch) = h.first_child(entry) {
+                h.set_parent(ch, cfg_node);
+            }
+            h.remove_node(entry);
+            let cfg_ty = h.optype_mut(cfg_node);
+            let OpType::CFG(CFG { signature }) = std::mem::take(cfg_ty) else {
+                unreachable!() // Checked at entry
+            };
+            let result_tys = signature.output.clone();
+            *cfg_ty = OpType::DFG(DFG { signature });
+            unpack_before_output(h, h.get_io(cfg_node).unwrap()[1], result_tys);
+            return Ok(NormalizeCFGResult::CFGToDFG);
+        } else if h.input_neighbours(entry).count() == 0 {
+            // 1b. Move entry block outside/before the CFG into a DFG; its successor becomes the entry block.
+            let new_cfg_inputs = entry_blk.successor_input(0).unwrap();
+            let dfg = h.add_node_with_parent(
+                cfg_parent,
+                DFG {
+                    signature: Signature::new(entry_blk.inputs.clone(), new_cfg_inputs.clone()),
+                },
+            );
+            let [_, entry_output] = h.get_io(entry).unwrap();
+            while let Some(n) = h.first_child(entry) {
+                h.set_parent(n, dfg);
+            }
+            h.move_before_sibling(succ, entry);
+            h.remove_node(entry);
+
+            unpack_before_output(h, entry_output, new_cfg_inputs.clone());
+
+            // Inputs to CFG go directly to DFG
+            for inp in h.node_inputs(cfg_node).collect::<Vec<_>>() {
+                for src in h.linked_outputs(cfg_node, inp).collect::<Vec<_>>() {
+                    h.connect(src.0, src.1, dfg, inp.index());
+                }
+                h.disconnect(cfg_node, inp);
+            }
+
+            // Update input ports
+            let OpType::CFG(cfg_ty) = h.optype_mut(cfg_node) else {
+                panic!()
+            };
+            let inputs_to_add =
+                new_cfg_inputs.len() as isize - cfg_ty.signature.input.len() as isize;
+            cfg_ty.signature.input = new_cfg_inputs;
+            h.add_ports(cfg_node, Direction::Incoming, inputs_to_add);
+
+            // Wire outputs of DFG directly to CFG
+            for src in h.node_outputs(dfg).collect::<Vec<_>>() {
+                h.connect(dfg, src, cfg_node, src.index());
+            }
+            entry_changed = true;
+        }
+    }
+    // 2. If the exit node has a single predecessor and that predecessor has no other successors...
+    let mut exit_changed = false;
+    if let Some(pred) = h
+        .input_neighbours(exit)
+        .exactly_one()
+        .ok()
+        .filter(|pred| h.output_neighbours(*pred).count() == 1)
+    {
+        // Code in that predecessor can be moved outside (into a new DFG after the CFG),
+        // and the predecessor deleted
+        let [_, output] = h.get_io(pred).unwrap();
+        let pred_blk = h.get_optype(pred).as_dataflow_block().unwrap();
+        let new_cfg_outs = pred_blk.inner_signature().into_owned().input;
+
+        // new CFG result type and exit block
+        let OpType::CFG(cfg_ty) = h.optype_mut(cfg_node) else {
+            panic!()
+        };
+        let result_tys = std::mem::replace(&mut cfg_ty.signature.output, new_cfg_outs.clone());
+        h.add_ports(
+            cfg_node,
+            Direction::Outgoing,
+            new_cfg_outs.len() as isize - result_tys.len() as isize,
+        );
+
+        *h.optype_mut(pred) = ExitBlock {
+            cfg_outputs: new_cfg_outs.clone(),
+        }
+        .into();
+        debug_assert_eq!(h.num_ports(pred, Direction::Outgoing), 1);
+        h.set_num_ports(pred, 1, 0);
+
+        h.move_before_sibling(pred, exit);
+        h.remove_node(exit);
+        // Move contents into new DFG
+        let dfg = h.add_node_with_parent(
+            cfg_parent,
+            DFG {
+                signature: Signature::new(new_cfg_outs, result_tys.clone()),
+            },
+        );
+        while let Some(n) = h.first_child(pred) {
+            h.set_parent(n, dfg);
+        }
+        unpack_before_output(h, output, result_tys);
+
+        // Move output edges.
+        for p in h.node_outputs(cfg_node).collect_vec() {
+            let tgts = h.linked_inputs(cfg_node, p).collect_vec();
+            h.disconnect(cfg_node, p);
+            for tgt in tgts {
+                h.connect(dfg, p, tgt.0, tgt.1)
+            }
+        }
+        for p in h.node_inputs(dfg).collect_vec() {
+            h.connect(cfg_node, p.index(), dfg, p);
+        }
+        exit_changed = true;
+    }
+    Ok(NormalizeCFGResult::CFGPreserved {
+        entry_changed,
+        exit_changed,
+    })
+}
+
+type NodePorts<N> = Vec<(N, OutgoingPort)>;
+/// Remove all input wires to `n` and return them in two groups:
+/// the [EdgeKind::Value] inputs, and the [EdgeKind::StateOrder] inputs
+fn take_inputs<H: HugrMut>(h: &mut H, n: H::Node) -> (NodePorts<H::Node>, NodePorts<H::Node>) {
+    let mut values = vec![];
+    let mut orders = vec![];
+    for p in h.node_inputs(n).collect_vec() {
+        let srcs = h.linked_outputs(n, p).collect_vec();
+        h.disconnect(n, p);
+        match h.get_optype(n).port_kind(p) {
+            Some(EdgeKind::Value(_)) => {
+                assert_eq!(srcs.len(), 1);
+                values.extend(srcs);
+            }
+            Some(EdgeKind::StateOrder) => {
+                assert_eq!(orders, []);
+                orders.extend(srcs);
+            }
+            k => panic!("Unexpected port kind: {:?}", k),
+        }
+    }
+    (values, orders)
+}
+
+fn tuple_elems<H: HugrView>(h: &H, n: H::Node, p: OutgoingPort) -> TypeRow {
+    match h.get_optype(n).port_kind(p) {
+        Some(EdgeKind::Value(ty)) => ty.as_sum().unwrap().as_tuple().unwrap().clone(),
+        p => panic!("Expected Value port not {:?}", p),
+    }
+    .try_into()
+    .unwrap()
+}
+
+/// Unpack the first `value_srcs`; wire the unpacked elements and remaining `value_srcs` into
+/// consecutive ports of `dst`. Finally wire `order_srcs` all to the order input of `dst`.
+pub(crate) fn wire_unpack_first<H: HugrMut>(
+    h: &mut H,
+    value_srcs: impl IntoIterator<Item = (H::Node, OutgoingPort)>,
+    order_srcs: impl IntoIterator<Item = (H::Node, OutgoingPort)>,
+    dst: H::Node,
+) {
+    let parent = h.get_parent(dst).unwrap();
+    let mut srcs = value_srcs.into_iter();
+    let src_to_unpack = srcs.next().unwrap();
+    let tuple_tys = tuple_elems(h, src_to_unpack.0, src_to_unpack.1);
+    let tuple_len = tuple_tys.len();
+    let unp = h.add_node_with_parent(parent, UnpackTuple::new(tuple_tys));
+    h.connect(src_to_unpack.0, src_to_unpack.1, unp, 0);
+
+    for i in 0..tuple_len {
+        h.connect(unp, i, dst, i);
+    }
+    assert_eq!(
+        h.get_optype(dst).other_port_kind(Direction::Incoming),
+        Some(EdgeKind::StateOrder)
+    );
+    let order_tgt = h.get_optype(dst).other_input_port().unwrap();
+    for (i, (src, src_p)) in srcs.enumerate() {
+        assert!(i + tuple_len < order_tgt.index());
+        h.connect(src, src_p, dst, i + tuple_len);
+    }
+    for (src, src_p) in order_srcs {
+        h.connect(src, src_p, dst, order_tgt);
+    }
+}
+
+/// Unpack the first input to specified [Output] node and shuffle all the rest along
+fn unpack_before_output<H: HugrMut>(h: &mut H, output_node: H::Node, new_types: TypeRow) {
+    let (values, orders) = take_inputs(h, output_node);
+    let OpType::Output(ou) = h.optype_mut(output_node) else {
+        panic!()
+    };
+    let ports_to_add = new_types.len() as isize - ou.types.len() as isize;
+    ou.types = new_types;
+    h.add_ports(output_node, Direction::Incoming, ports_to_add);
+    wire_unpack_first(h, values, orders, output_node);
+}
+
+#[cfg(test)]
+mod test {
+    use std::collections::{HashMap, HashSet};
+    use std::sync::Arc;
+
+    use hugr_core::extension::simple_op::MakeExtensionOp;
+    use itertools::Itertools;
+    use rstest::rstest;
+
+    #[expect(deprecated)]
+    // when merge_basic_blocks removed/hidden, consider testing normalize_cfgs directly
+    use super::{NormalizeCFGPass, NormalizeCFGResult, merge_basic_blocks, normalize_cfg};
+    use crate::ComposablePass;
+    use hugr_core::builder::{
+        CFGBuilder, Dataflow, HugrBuilder, SubContainer, endo_sig, inout_sig,
+    };
+    use hugr_core::extension::prelude::{ConstUsize, Noop, PRELUDE_ID, UnpackTuple, qb_t, usize_t};
+    use hugr_core::ops::{DataflowOpTrait, LoadConstant, OpTag, OpTrait, OpType, Tag};
+    use hugr_core::ops::{constant::Value, handle::NodeHandle};
+    use hugr_core::types::{Signature, Type, TypeRow};
+    use hugr_core::{Extension, HugrView, const_extension_ids, type_row};
+
+    const_extension_ids! {
+        const EXT_ID: ExtensionId = "TestExt";
+    }
+
+    fn extension() -> Arc<Extension> {
+        Extension::new_arc(
+            EXT_ID,
+            hugr_core::extension::Version::new(0, 1, 0),
+            |ext, extension_ref| {
+                ext.add_op(
+                    "Test".into(),
+                    String::new(),
+                    Signature::new(
+                        vec![qb_t(), usize_t()],
+                        TypeRow::from(vec![Type::new_sum(vec![vec![qb_t()], vec![usize_t()]])]),
+                    ),
+                    extension_ref,
+                )
+                .unwrap();
+            },
+        )
+    }
+
+    #[rstest]
+    #[case(true)]
+    #[case(false)]
+    fn merge_bbs_in_loop(#[case] self_loop: bool) -> Result<(), Box<dyn std::error::Error>> {
+        /* self_loop==False:
+           -> Noop1 -----> Test -> Exit       -> Noop1AndTest --> Exit
+               |            |            =>     /            \
+               \-<- Noop2 <-/                   \-<- Noop2 <-/
+           (Noop2 -> Noop1 cannot be merged because Noop1 is the entry node)
+
+           self_loop==True:
+           -> Noop --> Test -> Exit           -> NoopAndTest --> Exit
+               |        |                =>     /           \
+               \--<--<--/                       \--<-----<--/
+        */
+
+        let loop_variants: TypeRow = vec![qb_t()].into();
+        let exit_types: TypeRow = vec![usize_t()].into();
+        let e = extension();
+        let tst_op = e.instantiate_extension_op("Test", [])?;
+        let mut h = CFGBuilder::new(inout_sig(loop_variants.clone(), exit_types.clone()))?;
+        let mut no_b1 = h.simple_entry_builder(loop_variants.clone(), 1)?;
+        let n = no_b1.add_dataflow_op(Noop::new(qb_t()), no_b1.input_wires())?;
+        let br = no_b1.add_load_value(Value::unary_unit_sum());
+        let no_b1 = no_b1.finish_with_outputs(br, n.outputs())?;
+        let mut test_block = h.block_builder(
+            loop_variants.clone(),
+            vec![loop_variants.clone(), exit_types],
+            type_row![],
+        )?;
+        let [test_input] = test_block.input_wires_arr();
+        let usize_cst = test_block.add_load_value(ConstUsize::new(1));
+        let [tst] = test_block
+            .add_dataflow_op(tst_op, [test_input, usize_cst])?
+            .outputs_arr();
+        let test_block = test_block.finish_with_outputs(tst, [])?;
+        let loop_backedge_target = if self_loop {
+            no_b1
+        } else {
+            let mut no_b2 = h.simple_block_builder(endo_sig(loop_variants), 1)?;
+            let n = no_b2.add_dataflow_op(Noop::new(qb_t()), no_b2.input_wires())?;
+            let br = no_b2.add_load_value(Value::unary_unit_sum());
+            let nid = no_b2.finish_with_outputs(br, n.outputs())?;
+            h.branch(&nid, 0, &no_b1)?;
+            nid
+        };
+        h.branch(&no_b1, 0, &test_block)?;
+        h.branch(&test_block, 0, &loop_backedge_target)?;
+        h.branch(&test_block, 1, &h.exit_block())?;
+
+        let mut h = h.finish_hugr()?;
+        let r = h.entrypoint();
+        #[expect(deprecated)]
+        // when merge_basic_blocks removed/hidden, consider testing normalize_cfgs directly
+        merge_basic_blocks(&mut h);
+        h.validate().unwrap();
+        assert_eq!(r, h.entrypoint());
+        assert!(matches!(h.get_optype(r), OpType::CFG(_)));
+        let [entry, exit] = h
+            .children(r)
+            .take(2)
+            .collect::<Vec<_>>()
+            .try_into()
+            .unwrap();
+        // Check the Noop('s) is/are in the right block(s)
+        let nops = h
+            .entry_descendants()
+            .filter(|n| h.get_optype(*n).cast::<Noop>().is_some());
+        let (entry_nop, expected_backedge_target) = if self_loop {
+            assert_eq!(h.children(r).count(), 2);
+            (nops.exactly_one().ok().unwrap(), entry)
+        } else {
+            let [_, _, no_b2] = h.children(r).collect::<Vec<_>>().try_into().unwrap();
+            let mut nops = nops.collect::<Vec<_>>();
+            let entry_nop_idx = nops
+                .iter()
+                .position(|n| h.get_parent(*n) == Some(entry))
+                .unwrap();
+            let entry_nop = nops[entry_nop_idx];
+            nops.remove(entry_nop_idx);
+            let [n_op2] = nops.try_into().unwrap();
+            assert_eq!(h.get_parent(n_op2), Some(no_b2));
+            (entry_nop, no_b2)
+        };
+        assert_eq!(h.get_parent(entry_nop), Some(entry));
+        assert_eq!(
+            h.output_neighbours(entry).collect::<HashSet<_>>(),
+            HashSet::from([expected_backedge_target, exit])
+        );
+        // And the Noop in the entry block is consumed by the custom Test op
+        let tst = find_unique(
+            h.entry_descendants(),
+            |n| matches!(h.get_optype(*n), OpType::ExtensionOp(c) if c.def().extension_id() != &PRELUDE_ID),
+        );
+        assert_eq!(h.get_parent(tst), Some(entry));
+        assert_eq!(
+            h.output_neighbours(entry_nop).collect::<Vec<_>>(),
+            vec![tst]
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn elide_triple_with_permute() -> Result<(), Box<dyn std::error::Error>> {
+        // Blocks are just BB1 -> BB2 -> BB3 --> Exit.
+        // Should be merged into one BB (we don't check that specifically)
+        // and then the whole CFG elided.
+        let e = extension();
+        let tst_op: OpType = e.instantiate_extension_op("Test", &[])?.into();
+        let [res_t] = tst_op
+            .dataflow_signature()
+            .unwrap()
+            .into_owned()
+            .output
+            .into_owned()
+            .try_into()
+            .unwrap();
+        let mut h = CFGBuilder::new(inout_sig(qb_t(), res_t.clone()))?;
+        let mut bb1 = h.simple_entry_builder(vec![usize_t(), qb_t()].into(), 1)?;
+        let [inw] = bb1.input_wires_arr();
+        let load_cst = bb1.add_load_value(ConstUsize::new(1));
+        let pred = bb1.add_load_value(Value::unary_unit_sum());
+        let bb1 = bb1.finish_with_outputs(pred, [load_cst, inw])?;
+
+        let mut bb2 = h.block_builder(
+            vec![usize_t(), qb_t()].into(),
+            vec![type_row![]],
+            vec![qb_t(), usize_t()].into(),
+        )?;
+        let [u, q] = bb2.input_wires_arr();
+        let pred = bb2.add_load_value(Value::unary_unit_sum());
+        let bb2 = bb2.finish_with_outputs(pred, [q, u])?;
+
+        let mut bb3 = h.block_builder(
+            vec![qb_t(), usize_t()].into(),
+            vec![type_row![]],
+            res_t.clone().into(),
+        )?;
+        let [q, u] = bb3.input_wires_arr();
+        let tst = bb3.add_dataflow_op(tst_op, [q, u])?;
+        let pred = bb3.add_load_value(Value::unary_unit_sum());
+        let bb3 = bb3.finish_with_outputs(pred, tst.outputs())?;
+        // Now add control-flow edges between basic blocks
+        h.branch(&bb1, 0, &bb2)?;
+        h.branch(&bb2, 0, &bb3)?;
+        h.branch(&bb3, 0, &h.exit_block())?;
+
+        let mut h = h.finish_hugr()?;
+        let res = normalize_cfg(&mut h);
+        assert_eq!(res, Ok(NormalizeCFGResult::CFGToDFG));
+        h.validate()?;
+        assert_eq!(h.entrypoint_optype().tag(), OpTag::Dfg);
+        assert_eq!(
+            h.entry_descendants().find(|n| matches!(
+                h.get_optype(*n),
+                OpType::DataflowBlock(_) | OpType::CFG(_) | OpType::ExitBlock(_)
+            )),
+            None
+        );
+        let tst = find_unique(
+            h.entry_descendants(),
+            |n| matches!(h.get_optype(*n), OpType::ExtensionOp(c) if c.def().extension_id() != &PRELUDE_ID),
+        );
+        assert_eq!(h.get_parent(tst), Some(h.entrypoint()));
+
+        let inp = find_unique(h.entry_descendants(), |n| {
+            matches!(h.get_optype(*n), OpType::Input(_))
+        });
+        let mut tst_inputs = h.input_neighbours(tst).collect::<Vec<_>>();
+        tst_inputs.remove(tst_inputs.iter().find_position(|n| **n == inp).unwrap().0);
+        let [other_input] = tst_inputs.try_into().unwrap();
+        assert_eq!(
+            h.get_optype(other_input),
+            &(LoadConstant {
+                datatype: usize_t()
+            }
+            .into())
+        );
+        Ok(())
+    }
+
+    fn find_unique<T>(items: impl Iterator<Item = T>, pred: impl Fn(&T) -> bool) -> T {
+        items.filter(pred).exactly_one().ok().unwrap()
+    }
+
+    #[rstest]
+    fn elide_cfg() {
+        let ext = extension();
+        let op = ext.instantiate_extension_op("Test", []).unwrap();
+        let out_ty = op.signature().output().clone();
+        let mut cfg = CFGBuilder::new(op.signature().into_owned()).unwrap();
+        let mut entry = cfg.simple_entry_builder(out_ty, 1).unwrap();
+        let op_res = entry
+            .add_dataflow_op(op.clone(), entry.input_wires())
+            .unwrap();
+        let predicate = entry.add_load_value(Value::unary_unit_sum());
+        let entry = entry
+            .finish_with_outputs(predicate, op_res.outputs())
+            .unwrap();
+        cfg.branch(&entry, 0, &cfg.exit_block()).unwrap();
+
+        // Add an (unreachable) block before the entry
+        let Signature { input, output } = op.signature().as_ref().clone();
+        for (ty, dest) in [(input, entry), (output, cfg.exit_block())] {
+            let mut extra = cfg.simple_block_builder(endo_sig(ty), 1).unwrap();
+            let inp = extra.input_wires();
+            let branch = extra.add_load_value(Value::unary_unit_sum());
+            let extra = extra.finish_with_outputs(branch, inp).unwrap();
+            cfg.branch(&extra, 0, &dest).unwrap();
+        }
+
+        let mut h = cfg.finish_hugr().unwrap();
+
+        let func = h.children(h.module_root()).exactly_one().ok().unwrap();
+        assert_eq!(child_tags_ext_ids(&h, func), ["Input", "Output", "Cfg"]);
+        let mut dfb_children = child_tags_ext_ids(&h, entry.node());
+
+        let res = normalize_cfg(&mut h);
+        h.validate().unwrap();
+        assert_eq!(res, Ok(NormalizeCFGResult::CFGToDFG));
+        assert_eq!(h.entrypoint_optype().tag(), OpTag::Dfg);
+        assert_eq!(child_tags_ext_ids(&h, func), ["Input", "Output", "Dfg"]);
+        dfb_children.push("UnpackTuple".to_string());
+        assert_eq!(child_tags_ext_ids(&h, h.entrypoint()), dfb_children);
+    }
+
+    #[test]
+    fn entry_before_loop() -> Result<(), Box<dyn std::error::Error>> {
+        /* -> Noop --> Test -> Exit      -> Test --> Exit
+                       |  |          =>     |  |
+                       \<-/                 \<-/
+        */
+        let loop_variants: TypeRow = vec![qb_t()].into();
+        let exit_types: TypeRow = vec![usize_t()].into();
+        let e = extension();
+        let tst_op = e.instantiate_extension_op("Test", [])?;
+        let mut h = CFGBuilder::new(inout_sig(qb_t(), usize_t()))?;
+        let mut nop_b = h.simple_entry_builder(loop_variants.clone(), 1)?;
+        let n = nop_b.add_dataflow_op(Noop::new(qb_t()), nop_b.input_wires())?;
+        let br = nop_b.add_load_value(Value::unary_unit_sum());
+        let entry = nop_b.finish_with_outputs(br, n.outputs())?;
+
+        let mut loop_b = h.block_builder(
+            loop_variants.clone(),
+            [loop_variants, exit_types],
+            type_row![],
+        )?;
+        let [qb] = loop_b.input_wires_arr();
+        let usz = loop_b.add_load_value(ConstUsize::new(3));
+        let [tst] = loop_b.add_dataflow_op(tst_op, [qb, usz])?.outputs_arr();
+        let loop_ = loop_b.finish_with_outputs(tst, [])?;
+        h.branch(&entry, 0, &loop_)?;
+        h.branch(&loop_, 0, &loop_)?;
+        h.branch(&loop_, 1, &h.exit_block())?;
+
+        let mut h = h.finish_hugr()?;
+
+        let res = normalize_cfg(&mut h).unwrap();
+        h.validate().unwrap();
+        assert_eq!(
+            res,
+            NormalizeCFGResult::CFGPreserved {
+                entry_changed: true,
+                exit_changed: false
+            }
+        );
+        assert_eq!(
+            h.children(h.entrypoint())
+                .map(|n| h.get_optype(n).tag())
+                .collect_vec(),
+            [OpTag::DataflowBlock, OpTag::BasicBlockExit]
+        );
+        let func = h.get_parent(h.entrypoint()).unwrap();
+        let func_children = child_tags_ext_ids(&h, func);
+        assert_eq!(
+            func_children.into_iter().sorted().collect_vec(),
+            ["Cfg", "Dfg", "Input", "Output",]
+        );
+        let [dfg] = h
+            .children(func)
+            .filter(|n| h.get_optype(*n).is_dfg())
+            .collect_array()
+            .unwrap();
+        assert_eq!(
+            child_tags_ext_ids(&h, dfg)
+                .into_iter()
+                .sorted()
+                .collect_vec(),
+            [
+                "Const",
+                "Input",
+                "LoadConst",
+                "Noop",
+                "Output",
+                "UnpackTuple"
+            ]
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn loop_before_exit() -> Result<(), Box<dyn std::error::Error>> {
+        /* -> Test -> Noop -> Exit      -> Test --> Exit (then Noop)
+              |  |                  =>    |  |
+              \<-/                        \<-/
+        */
+        let loop_variants: TypeRow = vec![qb_t()].into();
+        let exit_types: TypeRow = vec![usize_t()].into();
+        let e = extension();
+        let tst_op = e.instantiate_extension_op("Test", [])?;
+        let mut h = CFGBuilder::new(inout_sig(loop_variants.clone(), exit_types.clone()))?;
+
+        let mut loop_b = h.entry_builder(vec![loop_variants, exit_types.clone()], type_row![])?;
+        let [qb] = loop_b.input_wires_arr();
+        let usz = loop_b.add_load_value(ConstUsize::new(3));
+        let [tst] = loop_b.add_dataflow_op(tst_op, [qb, usz])?.outputs_arr();
+        let loop_ = loop_b.finish_with_outputs(tst, [])?;
+        h.branch(&loop_, 0, &loop_)?;
+
+        let mut nop_b = h.simple_block_builder(endo_sig(exit_types), 1)?;
+        let n = nop_b.add_dataflow_op(Noop::new(usize_t()), nop_b.input_wires())?;
+        let br = nop_b.add_load_value(Value::unary_unit_sum());
+        let tail = nop_b.finish_with_outputs(br, n.outputs())?;
+
+        h.branch(&loop_, 1, &tail)?;
+        h.branch(&tail, 0, &h.exit_block())?;
+
+        let mut h = h.finish_hugr()?;
+        let res = normalize_cfg(&mut h).unwrap();
+        h.validate().unwrap();
+        assert_eq!(
+            res,
+            NormalizeCFGResult::CFGPreserved {
+                entry_changed: false,
+                exit_changed: true
+            }
+        );
+        assert_eq!(
+            h.children(h.entrypoint())
+                .map(|n| h.get_optype(n).tag())
+                .collect_vec(),
+            [OpTag::DataflowBlock, OpTag::BasicBlockExit]
+        );
+        let func = h.get_parent(h.entrypoint()).unwrap();
+        assert_eq!(
+            h.children(func)
+                .map(|n| h.get_optype(n).tag())
+                .collect_vec(),
+            [OpTag::Input, OpTag::Output, OpTag::Cfg, OpTag::Dfg]
+        );
+
+        let dfg = h.children(func).last().unwrap();
+        let dfg_children = child_tags_ext_ids(&h, dfg);
+        assert_eq!(
+            dfg_children.into_iter().sorted().collect_vec(),
+            [
+                "Const",
+                "Input",
+                "LoadConst",
+                "Noop",
+                "Output",
+                "UnpackTuple"
+            ]
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn nested_cfgs_pass() {
+        //  --> Entry --> Loop --> Tail --> EXIT
+        //        |       /  \
+        //      (E->X)    \<-/
+        let e = extension();
+        let tst_op = e.instantiate_extension_op("Test", []).unwrap();
+        let qqu = vec![qb_t(), qb_t(), usize_t()];
+        let qq = TypeRow::from(vec![qb_t(); 2]);
+        let mut outer = CFGBuilder::new(inout_sig(qqu.clone(), vec![usize_t(), qb_t()])).unwrap();
+        let mut entry = outer.entry_builder(vec![qq.clone()], type_row![]).unwrap();
+        let [q1, q2, u] = entry.input_wires_arr();
+        let (inner, inner_pred) = {
+            let mut inner = entry
+                .cfg_builder([(qb_t(), q1), (qb_t(), q2)], qq.clone())
+                .unwrap();
+            let mut entry = inner.entry_builder(vec![qq.clone()], type_row![]).unwrap();
+            let [q1, q2] = entry.input_wires_arr();
+            let [pred] = entry
+                .add_dataflow_op(Tag::new(0, vec![qq.clone()]), [q1, q2])
+                .unwrap()
+                .outputs_arr();
+            let entry = entry.finish_with_outputs(pred, []).unwrap();
+            inner.branch(&entry, 0, &inner.exit_block()).unwrap();
+            (inner.finish_sub_container().unwrap(), pred.node())
+        };
+        let [q1, q2] = inner.outputs_arr();
+        let [entry_pred] = entry
+            .add_dataflow_op(Tag::new(0, vec![qq.clone()]), [q1, q2])
+            .unwrap()
+            .outputs_arr();
+        let entry = entry.finish_with_outputs(entry_pred, []).unwrap();
+
+        let loop_b = {
+            let mut loop_b = outer
+                .block_builder(qq.clone(), [qb_t().into(), usize_t().into()], qb_t().into())
+                .unwrap();
+            let [q1, q2] = loop_b.input_wires_arr();
+            // u here is `dom` edge from entry block
+            let [pred] = loop_b
+                .add_dataflow_op(tst_op, [q1, u])
+                .unwrap()
+                .outputs_arr();
+            loop_b.finish_with_outputs(pred, [q2]).unwrap()
+        };
+        outer.branch(&entry, 0, &loop_b).unwrap();
+        outer.branch(&loop_b, 0, &loop_b).unwrap();
+
+        let (tail_b, tail_pred) = {
+            let uq = TypeRow::from(vec![usize_t(), qb_t()]);
+            let mut tail_b = outer
+                .block_builder(uq.clone(), vec![uq.clone()], type_row![])
+                .unwrap();
+            let [u, q] = tail_b.input_wires_arr();
+            let [br] = tail_b
+                .add_dataflow_op(Tag::new(0, vec![uq.clone()]), [u, q])
+                .unwrap()
+                .outputs_arr();
+            (tail_b.finish_with_outputs(br, []).unwrap(), br.node())
+        };
+        outer.branch(&loop_b, 1, &tail_b).unwrap();
+        outer.branch(&tail_b, 0, &outer.exit_block()).unwrap();
+        let mut h = outer.finish_hugr().unwrap();
+        assert_eq!(
+            h.get_parent(h.get_parent(inner_pred).unwrap()),
+            Some(inner.node())
+        );
+        assert_eq!(h.get_parent(entry_pred.node()), Some(entry.node()));
+        assert_eq!(h.get_parent(tail_pred.node()), Some(tail_b.node()));
+
+        let res = NormalizeCFGPass::default().run(&mut h).unwrap();
+        h.validate().unwrap();
+        assert_eq!(
+            res,
+            HashMap::from([
+                (inner.node(), NormalizeCFGResult::CFGToDFG),
+                (
+                    h.entrypoint(),
+                    NormalizeCFGResult::CFGPreserved {
+                        entry_changed: true,
+                        exit_changed: true
+                    }
+                )
+            ])
+        );
+        // Now contains only one CFG with one BB (self-loop)
+        assert_eq!(
+            h.nodes()
+                .filter(|n| h.get_optype(*n).is_cfg())
+                .exactly_one()
+                .ok(),
+            Some(h.entrypoint())
+        );
+        let [entry, exit] = h.children(h.entrypoint()).collect_array().unwrap();
+        assert_eq!(h.output_neighbours(entry).collect_vec(), [entry, exit]);
+        // Inner CFG is now a DFG (and still sibling of entry_pred)...
+        assert_eq!(h.get_parent(inner_pred), Some(inner.node()));
+        assert_eq!(h.get_optype(inner.node()).tag(), OpTag::Dfg);
+        assert_eq!(h.get_parent(inner.node()), h.get_parent(entry_pred.node()));
+        // Predicates lifted appropriately...
+        for n in [entry_pred.node(), tail_pred.node()] {
+            let parent = h.get_parent(n).unwrap();
+            assert_eq!(h.get_optype(parent).tag(), OpTag::Dfg);
+            assert_eq!(h.get_parent(parent), h.get_parent(h.entrypoint()));
+        }
+        // ...and followed by UnpackTuple's
+        for n in [inner_pred, entry_pred.node(), tail_pred.node()] {
+            let [unpack] = h.output_neighbours(n).collect_array().unwrap();
+            assert!(
+                h.get_optype(unpack)
+                    .as_extension_op()
+                    .and_then(|e| UnpackTuple::from_extension_op(e).ok())
+                    .is_some()
+            );
+        }
+    }
+
+    fn child_tags_ext_ids<H: HugrView>(h: &H, n: H::Node) -> Vec<String> {
+        h.children(n)
+            .map(|n| match h.get_optype(n) {
+                OpType::ExtensionOp(e) => e.unqualified_id().to_string(),
+                op => format!("{:?}", op.tag()),
+            })
+            .collect()
+    }
+}
